+++ src/data/instantiator.h
@@ -85,7 +85,7 @@ public:
 
     USigSet instantiate(const HtnOp& op, const std::function<bool(const Signature&)>& state);
     USigSet instantiateLimited(const HtnOp& op, const std::function<bool(const Signature&)>& state, 
-            const std::vector<int>& argsByPriority, int limit, bool returnUnfinished);
+            const std::vector<int>& argsByPriority, size_t limit, bool returnUnfinished);
 
     const FlatHashMap<int, float>& getPreconditionRatings(const USignature& opSig);
